getting there.  they now render

+++ common/pixelmon/entities/pixelmon/Entity4Textures.java
@@ -39,10 +39,10 @@ public abstract class Entity4Textures extends Entity3HasStats {
 	@SideOnly(Side.CLIENT)
 	@Override
 	public String getTexture() {
-		if (dataWatcher.getWatchableObjectShort(20) == 1
+		if (dataWatcher.getWatchableObjectShort(20) == (short)1
 				&& Minecraft.getMinecraft().renderEngine.texturePack.getSelectedTexturePack().getResourceAsStream("/pixelmon/texture/pokemon-shiny/shiny" + getName().toLowerCase() + ".png") != null)
 			return "/pixelmon/texture/pokemon-shiny/shiny" + getName().toLowerCase() + ".png";
-		else if (dataWatcher.getWatchableObjectShort(21) == 1
+		else if (dataWatcher.getWatchableObjectShort(21) == (short)1
 				&& Minecraft.getMinecraft().renderEngine.texturePack.getSelectedTexturePack().getResourceAsStream("/pixelmon/texture/pokemon-roasted/roasted" + getName().toLowerCase() + ".png") != null)
 			return "/pixelmon/texture/pokemon-roasted/roasted" + getName().toLowerCase() + ".png";
 		else
@@ -50,7 +50,7 @@ public abstract class Entity4Textures extends Entity3HasStats {
 	}
 
 	public boolean getIsShiny() {
-		return dataWatcher.getWatchableObjectShort(20) == 1;
+		return dataWatcher.getWatchableObjectShort(20) == (short)1;
 	}
 
 	public void setIsShiny(boolean isShiny) {
@@ -63,14 +63,14 @@ public abstract class Entity4Textures extends Entity3HasStats {
 	@Override
 	public void writeEntityToNBT(NBTTagCompound nbt) {
 		super.writeEntityToNBT(nbt);
-		nbt.setBoolean("IsShiny", dataWatcher.getWatchableObjectShort(20) == 1);
-		nbt.setBoolean("IsRoasted", dataWatcher.getWatchableObjectShort(21) == 1);
+		nbt.setBoolean("IsShiny", dataWatcher.getWatchableObjectShort(20) == (short)1);
+		nbt.setBoolean("IsRoasted", dataWatcher.getWatchableObjectShort(21) == (short)1);
 	}
 
 	@Override
 	public void readEntityFromNBT(NBTTagCompound nbt) {
 		super.readEntityFromNBT(nbt);
-		dataWatcher.updateObject(20, nbt.getBoolean("IsShiny") ? 1 : 0);
-		dataWatcher.updateObject(21, nbt.getBoolean("IsRoasted") ? 1 : 0);
+		dataWatcher.updateObject(20, nbt.getBoolean("IsShiny") ? (short)1 : (short)0);
+		dataWatcher.updateObject(21, nbt.getBoolean("IsRoasted") ? (short)1 : (short)0);
 	}
 }
